back/requirements.txt
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
Pillow==8.2.0
pycryptodome==3.10.1
PyMySQL==1.0.2
qrcode==6.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
xmltodict==0.12.0
zipp==3.4.1
